💥 Remove deprecated signatures of `fc.float`/`fc.double`
Drop any legacy signatures of `fc.float`/`fc.double` as planned in #992.
Any signature marked as deprecated on `fc.float`/`fc.double` will be dropped by this PR.

Originally merged as #1495 for alpha versions.
