+++ src/arbitrary/_internals/helpers/QualifiedObjectConstraints.ts
@@ -126,7 +126,7 @@ function defaultValues(constraints: { size: SizeForArbitrary }): Arbitrary<unkno
   return [
     boolean(),
     maxSafeInteger(),
-    double({ next: true }),
+    double(),
     string(constraints),
     oneof(string(constraints), constant(null), constant(undefined)),
   ];
