+++ src/arbitrary/float.ts
@@ -1,121 +1,119 @@
-import { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
-import { floatNext, FloatNextConstraints } from './_next/floatNext';
 import { integer } from './integer';
-
-/** @internal */
-function next(n: number): Arbitrary<number> {
-  return integer(0, (1 << n) - 1);
-}
-
-/** @internal */
-const floatInternal = (): Arbitrary<number> => {
-  // uniformaly in the range 0 (inc.), 1 (exc.)
-  return next(24).map((v) => v / (1 << 24));
-};
+import { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
+import { floatToIndex, indexToFloat, MAX_VALUE_32 } from './_internals/helpers/FloatHelpers';
+import { convertFromNext, convertToNext } from '../check/arbitrary/definition/Converters';
 
 /**
  * Constraints to be applied on {@link float}
  * @remarks Since 2.6.0
  * @public
  */
-export type FloatConstraints =
-  | {
-      /**
-       * Enable new version of fc.float
-       * @remarks Since 2.8.0
-       */
-      next?: false;
-      /**
-       * Lower bound for the generated floats (included)
-       * @remarks Since 2.6.0
-       */
-      min?: number;
-      /**
-       * Upper bound for the generated floats (excluded)
-       * @remarks Since 2.6.0
-       */
-      max?: number;
-    }
-  | ({
-      /**
-       * Enable new version of fc.float
-       * @remarks Since 2.8.0
-       */
-      next: true;
-    } & FloatNextConstraints);
+export interface FloatConstraints {
+  /**
+   * Lower bound for the generated 32-bit floats (included)
+   * @defaultValue Number.NEGATIVE_INFINITY, -3.4028234663852886e+38 when noDefaultInfinity is true
+   * @remarks Since 2.8.0
+   */
+  min?: number;
+  /**
+   * Upper bound for the generated 32-bit floats (included)
+   * @defaultValue Number.POSITIVE_INFINITY, 3.4028234663852886e+38 when noDefaultInfinity is true
+   * @remarks Since 2.8.0
+   */
+  max?: number;
+  /**
+   * By default, lower and upper bounds are -infinity and +infinity.
+   * By setting noDefaultInfinity to true, you move those defaults to minimal and maximal finite values.
+   * @defaultValue false
+   * @remarks Since 2.8.0
+   */
+  noDefaultInfinity?: boolean;
+  /**
+   * When set to true, no more Number.NaN can be generated.
+   * @defaultValue false
+   * @remarks Since 2.8.0
+   */
+  noNaN?: boolean;
+}
 
 /**
- * For floating point numbers between 0.0 (included) and 1.0 (excluded) - accuracy of `1 / 2**24`
+ * Same as {@link floatToIndex} except it throws in case of invalid float 32
+ *
+ * @internal
+ */
+function safeFloatToIndex(f: number, constraintsLabel: keyof FloatConstraints) {
+  const conversionTrick = 'you can convert any double to a 32-bit float by using `new Float32Array([myDouble])[0]`';
+  const errorMessage = 'fc.float constraints.' + constraintsLabel + ' must be a 32-bit float - ' + conversionTrick;
+  if (Number.isNaN(f) || (Number.isFinite(f) && (f < -MAX_VALUE_32 || f > MAX_VALUE_32))) {
+    // Number.NaN does not have any associated index in the current implementation
+    // Finite values outside of the 32-bit range for floats cannot be 32-bit floats
+    throw new Error(errorMessage);
+  }
+  const index = floatToIndex(f);
+  if (!Number.isInteger(index)) {
+    // Index not being an integer means that original value was not a valid 32-bit float
+    throw new Error(errorMessage);
+  }
+  return index;
+}
+
+/** @internal */
+function unmapperFloatToIndex(value: unknown): number {
+  if (typeof value !== 'number') throw new Error('Unsupported type');
+  return floatToIndex(value);
+}
+
+/**
+ * For 32-bit floating point numbers:
+ * - sign: 1 bit
+ * - significand: 23 bits
+ * - exponent: 8 bits
+ *
+ * The smallest non-zero value (in absolute value) that can be represented by such float is: 2 ** -126 * 2 ** -23.
+ * And the largest one is: 2 ** 127 * (1 + (2 ** 23 - 1) / 2 ** 23).
+ *
+ * @param constraints - Constraints to apply when building instances (since 2.8.0)
+ *
  * @remarks Since 0.0.6
  * @public
  */
-function float(): Arbitrary<number>;
-/**
- * For floating point numbers between 0.0 (included) and max (excluded) - accuracy of `max / 2**24`
- *
- * @param max - Upper bound of the generated floating point
- *
- * @deprecated
- * Superceded by `fc.float({max})` - see {@link https://github.com/dubzzz/fast-check/issues/992 | #992}.
- * Ease the migration with {@link https://github.com/dubzzz/fast-check/tree/main/codemods/unify-signatures | our codemod script}.
- *
- * @remarks Since 1.0.0
- * @public
- */
-function float(max: number): Arbitrary<number>;
-/**
- * For floating point numbers between min (included) and max (excluded) - accuracy of `(max - min) / 2**24`
- *
- * @param min - Lower bound of the generated floating point
- * @param max - Upper bound of the generated floating point
- *
- * @remarks You may prefer to use `fc.float({min, max})` instead.
- * @remarks Since 1.0.0
- * @public
- */
-function float(min: number, max: number): Arbitrary<number>;
-/**
- * For floating point numbers in range defined by constraints - accuracy of `(max - min) / 2**24`
- *
- * @param constraints - Constraints to apply when building instances
- *
- * @remarks Since 2.6.0
- * @public
- */
-function float(constraints: FloatConstraints): Arbitrary<number>;
-function float(...args: [] | [number] | [number, number] | [FloatConstraints]): Arbitrary<number> {
-  if (typeof args[0] === 'object') {
-    if (args[0].next) {
-      return floatNext(args[0]);
-    }
-    const min = args[0].min !== undefined ? args[0].min : 0;
-    const max = args[0].max !== undefined ? args[0].max : 1;
-    return (
-      floatInternal()
-        .map((v) => min + v * (max - min))
-        // For v = 0.9999999403953552, min: 0, max: 5e-324
-        // we have `min + v * (max - min)` equal to `max`
-        .filter((g) => g !== max || g === min)
-    );
-  } else {
-    const a = args[0];
-    const b = args[1];
-    if (a === undefined) return floatInternal();
-    if (b === undefined)
-      return (
-        floatInternal()
-          .map((v) => v * a)
-          // For v = 0.9999999403953552, a: 5e-324
-          // we have `v * a` equal to `a`
-          .filter((g) => g !== a || g === 0)
-      );
-    return (
-      floatInternal()
-        .map((v) => a + v * (b - a))
-        // For v = 0.9999999403953552, a: 0, b: 5e-324
-        // we have `a + v * (b - a)` equal to `b`
-        .filter((g) => g !== b || g === a)
+export function float(constraints: FloatConstraints = {}): Arbitrary<number> {
+  const {
+    noDefaultInfinity = false,
+    noNaN = false,
+    min = noDefaultInfinity ? -MAX_VALUE_32 : Number.NEGATIVE_INFINITY,
+    max = noDefaultInfinity ? MAX_VALUE_32 : Number.POSITIVE_INFINITY,
+  } = constraints;
+  const minIndex = safeFloatToIndex(min, 'min');
+  const maxIndex = safeFloatToIndex(max, 'max');
+  if (minIndex > maxIndex) {
+    // Comparing min and max might be problematic in case min=+0 and max=-0
+    // For that reason, we prefer to compare computed index to be safer
+    throw new Error('fc.float constraints.min must be smaller or equal to constraints.max');
+  }
+  if (noNaN) {
+    return convertFromNext(
+      convertToNext(integer({ min: minIndex, max: maxIndex })).map(indexToFloat, unmapperFloatToIndex)
     );
   }
+  // In case maxIndex > 0 or in other words max > 0,
+  //   values will be [min, ..., +0, ..., max, NaN]
+  //               or [min, ..., max, NaN] if min > +0
+  // Otherwise,
+  //   values will be [NaN, min, ..., max] with max <= +0
+  const minIndexWithNaN = maxIndex > 0 ? minIndex : minIndex - 1;
+  const maxIndexWithNaN = maxIndex > 0 ? maxIndex + 1 : maxIndex;
+  return convertFromNext(
+    convertToNext(integer({ min: minIndexWithNaN, max: maxIndexWithNaN })).map(
+      (index) => {
+        if (index > maxIndex || index < minIndex) return Number.NaN;
+        else return indexToFloat(index);
+      },
+      (value) => {
+        if (typeof value !== 'number') throw new Error('Unsupported type');
+        if (Number.isNaN(value)) return maxIndex !== maxIndexWithNaN ? maxIndexWithNaN : minIndexWithNaN;
+        return floatToIndex(value);
+      }
+    )
+  );
 }
-export { float };
