+++ test/e2e/NoRegression.spec.ts
@@ -68,26 +68,10 @@ describe(`NoRegression`, () => {
       )
     ).toThrowErrorMatchingSnapshot();
   });
-  it('floatNext', () => {
-    expect(() =>
-      fc.assert(
-        fc.property(fc.float({ next: true }), (v) => testFunc(v)),
-        settings
-      )
-    ).toThrowErrorMatchingSnapshot();
-  });
   it('double', () => {
     expect(() =>
       fc.assert(
         fc.property(fc.double(), (v) => testFunc(v)),
-        settings
-      )
-    ).toThrowErrorMatchingSnapshot();
-  });
-  it('doubleNext', () => {
-    expect(() =>
-      fc.assert(
-        fc.property(fc.double({ next: true }), (v) => testFunc(v)),
         settings
       )
     ).toThrowErrorMatchingSnapshot();
