+++ test/unit/arbitrary/float32Array.spec.ts
@@ -24,8 +24,8 @@ describe('float32Array (integration)', () => {
       {
         minLength: fc.nat({ max: 5 }),
         maxLength: fc.nat({ max: 25 }),
-        min: fc.float({ next: true, noDefaultInfinity: true, noNaN: true }),
-        max: fc.float({ next: true, noDefaultInfinity: true, noNaN: true }),
+        min: fc.float({ noDefaultInfinity: true, noNaN: true }),
+        max: fc.float({ noDefaultInfinity: true, noNaN: true }),
         noDefaultInfinity: fc.boolean(),
         noNaN: fc.boolean(),
       },
