+++ test/unit/utils/stringify.spec.ts
@@ -60,7 +60,19 @@ describe('stringify', () => {
   it('Should be equivalent to JSON.stringify for JSON compliant objects', () =>
     fc.assert(
       fc.property(
-        fc.anything({ values: [fc.boolean(), fc.integer(), fc.double(), fc.fullUnicodeString(), fc.constant(null)] }),
+        // Remark: While fc.unicodeJsonObject() could have been a good alternative to fc.anything()
+        //         it unfortunately cannot be used as JSON.stringify poorly handles negative zeros.
+        // JSON.parse('{"a": -0}') -> preserves -0
+        // JSON.stringify({a: -0}) -> changes -0 into 0, it produces {"a":0}
+        fc.anything({
+          values: [
+            fc.boolean(),
+            fc.integer(),
+            fc.double({ noDefaultInfinity: true, noNaN: true }).filter((d) => !Object.is(d, -0)),
+            fc.fullUnicodeString(),
+            fc.constant(null),
+          ],
+        }),
         (obj) => {
           expect(stringify(obj)).toEqual(JSON.stringify(obj));
         }
